+++ src/main/java/com/cognite/client/Timeseries.java
@@ -19,11 +19,11 @@ package com.cognite.client;
 import com.cognite.client.config.ResourceType;
 import com.cognite.client.config.UpsertMode;
 import com.cognite.client.dto.Aggregate;
-import com.cognite.client.dto.Event;
 import com.cognite.client.dto.TimeseriesMetadata;
 import com.cognite.client.dto.Item;
 import com.cognite.client.servicesV1.ConnectorServiceV1;
 import com.cognite.client.servicesV1.parser.TimeseriesParser;
+import com.cognite.client.util.Items;
 import com.google.auto.value.AutoValue;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -120,7 +120,29 @@ public abstract class Timeseries extends ApiBase {
     }
 
     /**
-     * Retrieves timeseries by id.
+     * Retrieve timeseries by {@code externalId}.
+     *
+     * @param externalId The {@code externalIds} to retrieve
+     * @return The retrieved timeseries.
+     * @throws Exception
+     */
+    public List<TimeseriesMetadata> retrieve(String... externalId) throws Exception {
+        return retrieve(Items.parseItems(externalId));
+    }
+
+    /**
+     * Retrieve timeseries by {@code internal id}.
+     *
+     * @param id The {@code ids} to retrieve
+     * @return The retrieved timeseries.
+     * @throws Exception
+     */
+    public List<TimeseriesMetadata> retrieve(long... id) throws Exception {
+        return retrieve(Items.parseItems(id));
+    }
+
+    /**
+     * Retrieves timeseries by {@code externalId / id}.
      *
      * @param items The item(s) {@code externalId / id} to retrieve.
      * @return The retrieved timeseries.