+++ releases.md
@@ -19,7 +19,7 @@ Changes are grouped as follows:
 
 - Geo-location attribute and resource type.
 
-## [1.10.0-SNAPSHOT]
+## [1.10.0] 2022-02-02
 
 ### Added
 
@@ -28,10 +28,18 @@ Changes are grouped as follows:
 - Added `3D Asset Mapping`
 - `EngineeringDiagrams` promoted from experimental to stable. It has the same signature and behavior as before and is 
 located under the `contextualization` family: `CogniteClient.contextualization().engineeringDiagrams()`.
+- Added convenience methods to the `Request` object for easier handling of items (by `externalId` or `id`). You can use 
+`Request.withItemExternalIds(String... externalId)` and `Request.withItemInternalIds(Long... externalId)` to add multiple 
+items to the request.
+- Added convenience methods for retrieving items by `externalId` and `id`: `client.<resourceType>().retrieve(String... externalId)` 
+and `client.<resourceType>().retrieve(String... externalId)`. This is implemented by the resource types `Assets`, `DataPoints`, 
+`Datasets`, `Events`, `ExtractionPipelines`, `Files`, `Relationships`, `Sequences` and `SequencesRows`.
 
 ### Deprecated
 
 - The experimental version of `EngineeringDiagrams` is deprecated given the new, stable version.
+- The single item methods `Request.withItemExternalId(String externalId)` and `Request.withItemInternalId(Long externalId)` 
+have been deprecated in favour of the new multi-item versions.
 
 ### Removed